project/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
boto3==1.26.11
botocore==1.29.11
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.16
django-allauth==0.51.0
django-appconf==1.0.5
django-bootstrap5==22.1
django-braces==1.15.0
django-extensions==3.2.1
django-imagekit==4.1.0
django-model-utils==4.2.0
django-redis==5.2.0
django-social-share==2.3.0
django-star-ratings==0.9.2
django-storages==1.13.1
django-summernote==0.8.20.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
docker==6.0.1
et-xmlfile==1.1.0
exceptiongroup==1.0.1
executing==1.1.1
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.1.0
incremental==22.10.0
ipython==8.5.0
jedi==0.18.1
jmespath==1.0.1
lorem==0.1.1
Markdown==3.4.1
matplotlib-inline==0.1.6
msgpack==1.0.4
mypy-extensions==0.4.3
numpy==1.23.4
oauthlib==3.2.2
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
pywin32==306
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
selenium==4.6.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.1
swapper==1.3.0
tomli==2.0.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0
zipp==3.11.0
zope.interface==5.5.1
